+++ client/src/game/input/keyboard.ts
@@ -1,4 +1,4 @@
-import { Vector } from "@/game/geom";
+import { GlobalPoint, Vector } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
 import { copyShapes, deleteShapes, pasteShapes } from "@/game/shapes/utils";
 import { DEFAULT_GRID_SIZE, gameStore } from "@/game/store";
@@ -11,6 +11,7 @@ import { EventBus } from "../event-bus";
 import { floorStore } from "../layers/store";
 import { moveFloor } from "../layers/utils";
 import { gameManager } from "../manager";
+import { gameSettingsStore } from "../settings";
 
 export function onKeyUp(event: KeyboardEvent): void {
     if (event.target instanceof HTMLInputElement || event.target instanceof HTMLTextAreaElement) {
@@ -19,7 +20,9 @@ export function onKeyUp(event: KeyboardEvent): void {
         if (event.key === "Delete" || event.key === "Del" || event.key === "Backspace") {
             deleteShapes();
         }
-        if (event.key === " ") {
+        if (event.key === " " || (event.code === "Numpad0" && !event.ctrlKey)) {
+            // Spacebar or numpad-zero: cycle through own tokens
+            // numpad-zero only if Ctrl is not pressed, as this would otherwise conflict with Ctrl + 0
             const tokens = gameStore.ownedtokens.map(o => layerManager.UUIDMap.get(o)!);
             if (tokens.length === 0) return;
             const i = tokens.findIndex(o => o.center().equals(gameStore.screenCenter));
@@ -41,16 +44,60 @@ export function onKeyDown(event: KeyboardEvent): void {
         // Ctrl-a with a HTMLInputElement or a HTMLTextAreaElement selected - select all the text
         if (event.key === "a" && event.ctrlKey) event.target.select();
     } else {
-        if (event.keyCode >= 37 && event.keyCode <= 40) {
-            // Arrow keys - move the selection or the camera
+        const navKeys = [
+            "ArrowLeft",
+            "ArrowRight",
+            "ArrowUp",
+            "ArrowDown",
+            "Numpad1",
+            "Numpad2",
+            "Numpad3",
+            "Numpad4",
+            "Numpad6",
+            "Numpad7",
+            "Numpad8",
+            "Numpad9",
+        ];
+        if (navKeys.includes(event.code)) {
+            // Arrow keys & Numpad 1-4/6-9 - move the selection or the camera
             // todo: this should already be rounded
             const gridSize = DEFAULT_GRID_SIZE;
-            let offsetX = gridSize * (event.keyCode % 2);
-            let offsetY = gridSize * (event.keyCode % 2 ? 0 : 1);
+            let offsetX = 0;
+            let offsetY = 0;
+            const left = ["ArrowLeft", "Numpad1", "Numpad4", "Numpad7"];
+            const right = ["ArrowRight", "Numpad3", "Numpad6", "Numpad9"];
+            const up = ["ArrowUp", "Numpad7", "Numpad8", "Numpad9"];
+            const down = ["ArrowDown", "Numpad1", "Numpad2", "Numpad3"];
+            if (left.includes(event.code)) {
+                offsetX -= gridSize;
+            }
+            if (right.includes(event.code)) {
+                offsetX += gridSize;
+            }
+            if (up.includes(event.code)) {
+                offsetY -= gridSize;
+            }
+            if (down.includes(event.code)) {
+                offsetY += gridSize;
+            }
+            // in hex mode, if movement is diagonal, offsets have to be modified
+            if (gameSettingsStore.gridType === "FLAT_HEX" && offsetX != 0) {
+                offsetX = (1.5 * offsetX) / Math.sqrt(3);
+                offsetY *= 0.5;
+            } else if (gameSettingsStore.gridType === "POINTY_HEX" && offsetY != 0) {
+                offsetX *= 0.5;
+                offsetY = (1.5 * offsetY) / Math.sqrt(3);
+            }
             if (layerManager.hasSelection()) {
+                // if in hex-mode, first invalidate invalid axis
+                const flatHexInvalid = ["ArrowLeft", "ArrowRight", "Numpad4", "Numpad6"];
+                const pointyHexInvalid = ["ArrowUp", "ArrowDown", "Numpad2", "Numpad8"];
+                if (gameSettingsStore.gridType === "FLAT_HEX" && flatHexInvalid.includes(event.code)) {
+                    offsetX = 0;
+                } else if (gameSettingsStore.gridType === "POINTY_HEX" && pointyHexInvalid.includes(event.code)) {
+                    offsetY = 0;
+                }
                 const selection = layerManager.getSelection();
-                offsetX *= event.keyCode <= 38 ? -1 : 1;
-                offsetY *= event.keyCode <= 38 ? -1 : 1;
                 let delta = new Vector(offsetX, offsetY);
                 if (!event.shiftKey || !gameStore.IS_DM) {
                     // First check for collisions.  Using the smooth wall slide collision check used on mouse move is overkill here.
@@ -95,8 +142,8 @@ export function onKeyDown(event: KeyboardEvent): void {
                 floorStore.currentLayer!.invalidate(false);
             } else {
                 // The pan offsets should be in the opposite direction to give the correct feel.
-                gameStore.increasePanX(offsetX * (event.keyCode <= 38 ? 1 : -1));
-                gameStore.increasePanY(offsetY * (event.keyCode <= 38 ? 1 : -1));
+                gameStore.increasePanX(offsetX * -1);
+                gameStore.increasePanY(offsetY * -1);
                 layerManager.invalidateAllFloors();
                 sendClientLocationOptions();
             }
@@ -116,9 +163,24 @@ export function onKeyDown(event: KeyboardEvent): void {
             event.stopPropagation();
             gameStore.toggleUI();
         } else if (event.key === "0" && event.ctrlKey) {
-            // Ctrl-0 - Re-center/reset the viewport
-            gameStore.setPanX(0);
-            gameStore.setPanY(0);
+            // Ctrl-0 or numpad 0 - Re-center/reset the viewport
+            gameManager.setCenterPosition(new GlobalPoint(0, 0));
+            sendClientLocationOptions();
+            layerManager.invalidateAllFloors();
+        } else if (event.code === "Numpad5") {
+            // numpad 5 will center on selected shape(s) or on origin
+            let targetX = 0;
+            let targetY = 0;
+            if (layerManager.hasSelection()) {
+                const selection = layerManager.getSelection();
+                for (const sel of selection) {
+                    targetX += sel.refPoint.x;
+                    targetY += sel.refPoint.y;
+                }
+                targetX /= selection.length;
+                targetY /= selection.length;
+            }
+            gameManager.setCenterPosition(new GlobalPoint(targetX, targetY));
             sendClientLocationOptions();
             layerManager.invalidateAllFloors();
         } else if (event.key === "c" && event.ctrlKey) {